move tests into test projects
